level work
Got the basis for a level system down.
Started with a 2D array system, but I decided against using walls as a gameplay system for now.
I decided on a semi-3D array system where lists of enemy attributes are stored in lists of enemies which are stored in a list of levels.

To Do:
Add different Launcher Variations