Polish and final release
Barring any major bugs, this will probably be the last push.

Added a tutorial. The code is pretty messy but it works

Made a final level. (now 26 in total)

Added a title screen

Removed sound and sprites

Added new Ghost Mode if the player presses G in the pause menu or beats the final level

I would have kept images and sound but I wanted to prioritize a polished game.